created seeder and faker

+++ Demo/database/migrations/2014_10_12_000000_create_users_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
-            $table->integer('phone')->nullable();
+            $table->bigInteger('phone')->nullable();
             $table->boolean('isActive')->default(1);
             $table->softDeletes();
             $table->rememberToken();
